From Harro,
Hi all,

I tracked down an important bug in MSFitsInput which prevented users
from reading their UVFITS data into a MeasurementSet. Even though the
format is antiquated, some apps still do use/produce it.

It's a simple thing, an uninitialized variable which causes a huge
amount of memory to be (tried to) allocated which is denied by throwing
a std::bad_alloc exception.

This bug is (still) present in the "AIPS++ version: 1.9 (build #1574)"

The patch is made against the MSFitsInput.cc file from this version.

Cheers,

harro