put larger limit on the sort parallelization and limit worker group to 2
Even with the previously applied limit there one can still see
measurable performance degradations due to the poor parallelization of
the quicksort.
Try to reduce them some more by increasing the threshold and reducing
the worker group to the 2 threads the code is currently using.
This should remove the overhead due to the barrier.