Cube Tile Reuse
Previously whenver the TSMCube cube object needed to evict a tile from
cache and add a new one it would write the target tile, if needed, and
then deallocate it; then a new tile was allocated and filled.  This
mod caches a tile whenver it is deallocated and allows the allocation
logic to use that cached tile rather than allocated a brand new one.
This approach reduces the amount of allocation which can become
significant when tiles are large (and they are likely to get larger in
HPC applications).

Ran the various tests named tTiled* using valgrind and fixed the one
potential error that was introduced.  Valgrind reports the writing of
unitialized storage but these errors predate this mod.