Introduce pread/pwrite calls to IO
pread/pwrite combines lseek + read/write into one system call but does
not change the offset of the file descriptor.
This is useful when multiple threads work on the same file descriptor.
Also most of casa's IO is actually of this type so this could be used to
cut the number of system calls per IO request in half. But as it does
not change the file descriptor actually using it needs some
investigation if code relies on the position of the descriptor.
For now only use it for the MultiFile IO as that is a new method that is
not directly used anywhere else except casacore.