This is to undo some merge problems that stepped on updates to the version in casacore/master.
CASA uses a somewhat unique numbering scheme for it's shared libraries. In this cmake file if the user sets CASA_BUILD=1 then the build logic will use CASA's numbering scheme; when that setting is not utilized, only the standard cmake logic will be invoked.

I'm making this pull request at the request of our build guy who's out of the country on an extended vacation (lucky bastard). The problem we have is that we build and semi-release lots of versions of the various shared libraries so we need a scheme to help manage them.