Modify MSIter to allow using memory resident sort tables
Added a feature to modify MSIter so that the user has the option of specifying
that MSIter use a memory-resident table for the SORT_TABLE that it creates.
In concurrent operation a disk-resident SORT_TABLE allows two readonly
processes to interfere via the sort table.  When the new, optional ctor
parameter storeSort is set to false MSIter will use a memory resident
sort table that will vanish with the end of the creating process.  The
default value of the new parameter is true which means that the operation
of the ctor will be the same for existing code.