SPW parser error handling improvements
Handling of errors in SPW parser is via an error handler now. This
brings uniform error reporting and allows changing the error reporting
behaviour programmatically by the clients.

Instead of throwing an exception for no-match for some components of
the expression, the error is now reported as a warning.  If no
components were matched, an exception is thrown.

Added new class MSSSpwErrorHandler.