Avoid heap allocations for small Unit strings
Small allocations are very expensive with glibc, if they are performed
often it can be very worthwhile to replace them with stack based
buffers.
Unit::check is a function which can be called very often by unoptimized
code. To reduce the impact of this avoid the heap allocation for small
strings.