Fix endian detection for ARM
This is a belt-and-braces fix: if __BYTE_ORDER__ is set (defined by GCC and
Clang), then that is used instead of trying to guess based on a list of known architectures. The architecture list is still there as a fallback, and I've added armel and aarch64el to the list of little-endian architectures.

I've also added aarch64 to the list of 64-bit architectures.

I've tested that this works on an armel system, which was previously unable to
read a measurement set due to the endianness being mis-detected. The aarch64
support is untested.