Add -fsigned-char to compiler arguments
casacore assumes in several places that char is signed, even though C/C++
does not guarantee it. Passing -fsigned-char forces the compiler to treat it as
signed even on architectures (such as ARM) where it is normally unsigned.

This is potentially risky if inlined header code in any included library
assumes that char is unsigned, but this change makes all the unit tests pass
so it is almost certainly better than not using it.

See #249.