merge unstrided tile copies
Often the tiles are copied to contigous or at least contigous subtiles
so the main copy loops in the unstrided tile access function can be
merged to avoid a lot of iterations and small copies.
E.g. CASA often uses (X, Y, Z) tiles with X being 2 or 4, so without
merging a series of 2 or 4 element copies are performed, this can now
often be merged to X*Y*Z sized copies.