Mod to allow setting tile cache using hypercube number instead of row.
The tile cache for an StMan is set on a per hypercube basis.  The methods
available from the TiledStManAccessor class use a row number that the
StMan then uses to determine the appropriate hypercube.  For virtually
concatenated databases, it's handy to be able to walk through the
hypercubes without regard to the row number.  TiledStManAccessor already
provides the number of tiles, so the new feature allows looping over the
hypercubes in the StMan and setting each hypercube to have the appropriate
cache size.

I don't fell that this mod does not appear to significantly expose any
additional abstractions from the StMan beyond what TiledStManAccessor
already exposes.