Thread-safety Aipsrc initialization.
- Rework code to use CallOnce.
Also recursive init here, but reworked one of the find() to findNoParse() to get rid of it.
Parts of the changed lines are reindentation from CallOnce.

- Document the PitA reRead().
This is one of those functions that is trivial to add. When multi-threading shows up, it's a real pain,
since now technically, any accessed aipsrc variable can change at any time.
Don't accomodate that flexibility: lot of effort and pointless to change config while processing.
Also document that the cached iau2000 and iau2000a vars in MeasTable are not re-inited on reRead().