Major refactoring of Measures Meas{IERS,JPL,Table} wrt thread-safe, fast, lazy init
Rework the code to always use CallOnce or CallOnce0 for lazy initialization.
This worked out by and large, but there's a lot of code move, reindentation and some cleanup in between...

But prefer to rely on C++11 compilers (and GCC, Clang and maybe others for pre-C++11) to init function scope
statics in a thread-safe way. This can even be slightly more efficient than std::call_once(). Need to rework a lot of code to make this possible.

For reviewers: so there is a common patterns you can review once if you don't want to go through 1000 lines changed...

In addition, at the end of MeasTable.cc:
- dUTC(): need to pack LEAP and N statics into a struct for safe init by construction via single return value.
- dUT1(): Esp. review this one. I added thread_local, 95% sure I understand what the intention of the code is. Then, thread_local is the way to go.
  If you have a better way to treat __APPLE__, let me know.

- msgDone cases: Mark the variable with C++11 atomic (otherwise undef behavior), but as was already indicated,
don't grab a lock or anything, because we don't care if we see a Measures warning message twice (or garbled).

Drop the volatile qualifiers. Volatile doesn't really help if accessing a single volatile var between lock/unlock.
It also doesn't help for the msgDone cases (we don't care anyway) and the compiler cannot remove the store,
since that'd violate single thread behavior on a 2nd invocation.

Misc:
- move the iau2000 and iau200a class statics to function local scope. Document that Aipsrc::reRead() does not update them.
- Some changes along: String("bla") -> "bla", as well as log lines and line break reflowing. Remove trailing whitespaces, also in a few static const tables.
- Fixed CASADATA. The default data dir changed a few months ago.

Compile-tested in various configs (but not __APPLE__). All regression tests pass (but I don't have SOFA for that IAU test case).