Set our CTest default timeout to 150 s
The default of 1500 is crazy.
A timeout of 150 s may be too low for valgrind tests.
With LOFAR, 150 s is plenty and LOFAR tests run on average and max way longer.

You can override it on the cmd-line with 'ctest --timeout xx'

(Note that changing values like done in this patch manually has a really odd,
delayed effect on what CTest does. I've seen it before and don't get it. But this patch works.)