consider cgroup limits for memory in HostInfo
You usually want to know the available memory that can be used when
using these functions but it returns all memory of a machine instead.
This causes problems when a process is resource limited via cgroups.

Update the function to consider the cgroup limits when returning its
results.
Currently only supports v1 cgroups, but currently there are no
standard distributions using v2 yet.

Currently hardcodes common cgroup mount point, technically one has to
search the location in the mount list.
It might be useful to change it to use libcgroup/cgget to be more
generic but the current layout works for ubuntu/debian/fedora/el7.

Closes gh-446