Remove table flush on destruction of PagedArray
The PagedArray destructor flushes its underlying Table which essentially
makes the write Tile cache useless as many short lived objects do
contain PagedArray objects. When iterating over a Lattice in an
order that does not match the tileshape it will flush the cache on
destructions of for example LatticeIterators, LatticeExpressions or
SubImages and it while thus constantly rewrite partially modified tiles.
In some CASA tasks this leads to unnecessary very large write
amplifications.

To remove the all code that explicitly looks at table files on disk must
be adapted to respect the TableCache as it the flush frequency is
massively reduced and the files might not exist on disk at the time the
functions try to read them.
This commit updates Table::isReadable and TableInfo which is enough to
make most of the CASA regression suite succeed.
This also requires removing an unnecessary lock from
TableCache::lookCache which is during the reopenRW call may otherwise
recursively lock the TableCache.

IsWritable is not changed as it does not appear necessary currently and
if would need additional logic to handle read-only tables in cache that
are supposed to be opened in write mode.

closes gh-453