Fix possible uInt overflows and tighten scopes
Several loops iterating arrays used uInt as counter, this could overflow
for large arrays.
Also minimize the scope of the accumulators in some of the computation
loops to avoid unnecessary stack spills. This changes the summation order,
though it will be better on average as it is more likely that the two summed
values have similar magnitude.