Make constant int set taql select performance consistent
Replace the limited to 1024**2 constant lookup table with a set making
the performance logarithmic for all sizes.
Avoids unexpectedly degrading to quadratic performance on large ranges.
The constant time lookup is not relevant enough to keep the extra code,
if required using a c++11 unorderd_set is probably the better choice.