Don't build tests by default
The autotools behaviour of building tests only on the make check target
is much better as it avoids having to go through the length test
compilation and uncacheable linking during development.
Try to emulate this behaviour with cmake by adding a check target and
only building the test directories on this target.
As you can't add test dependencies on the test target it is
disfunctional when check has not been executed first.

For backward compatibility when BUILD_TESTING is set everything remains
as is, so make builds the test and make test runs them.