Support keywords in python conversion
Previously, conversion to/from TableDesc keywords was not supported.
This is now supported at both table and column level. The python
dictionary associated with a column now contains a "keywords" key, while
the table dictionary now contains both a "_keywords_". A
"_private_keywords_" key also now exists but is only created when
converting from TableDesc to python dictionary as these objects are
private on the TableDesc object.