Remove use of assert()
CASA Assert* macro definitions are dependent on whether NDEBUG is defined or
not, such that when NDEBUG is defined, the CASA macros are defined in terms of
the standard assert() macro. However, when NDEBUG is defined, the assert() macro
is effectively a no-op. The drawback of maintaining the inclusion of the
assert() macro is that it requires the assert.h header file, which is actually
missing from the Error.h file. Rather than include the header file so that
compilation works in the case of NDEBUG being defined, removing the use of
assert() in that case, in which it becomes a no-op, is the better solution.