Fix FTBFS tStatisticsUtilities, tLatticeStatistics and tLCEllipsoid on arm64
The tests "tStatisticsUtilities" and "tLatticeStatistics" can be made
to pass on arm64 with these adjustments to the expected accuracy.

The test "tLCEllipsoid" seems to be converting ellipses into bitmaps.
I'm guessing it gives different output on amd64 and arm64 because the
real curve passes through a critical point and numerical inaccuracy
puts it on different sides of the point on the two architectures. One
way to fix that might be to use non-round numbers in the parameters so
that the curve does not pass through a critical point. That's what I
tried to do, and with the following changes to the program and the
expected output the test passes on both amd64 and arm64.