Check for not null DATA or FLOAT DATA columns before accessing the keywords
The setMSInfo() function would check the keywords of the DATA column,
or if it doesn't exist FLOAT DATA in order to get a possible channel selection.
If you create a synthethic MS with MsFactory (from CASA) without any of the
two columns a segfault happens. Still, such an MS can be useful for unit
testing purposes.