Unify calls to wcsini and wcssub
Unification is necessary not only to avoid duplication, but because we
need to serialize these calls, and it's easier to do it if we use them
centrally. In wcslib >= 5.7 [1] wcssub internally calls wcssnpv/wcssnps
to temporarily set two global variables to a particular value before
calling wcsini, which in turn reads their values; after wcsini returns
wcssub sets the variables to their previous values. This situation
creates a race condition between concurrent, independent calls to wcssub
and wcsini. Thus, we serialize them with a lock.

The issue has been notified to Mark Calabretta, so hopefully we'll see a
fix for this issue in newer wcslib versions. For the time being we apply
this fix to all wcslib versions >= 5.7, but after a new, fixed version of
wcslib is available we can constrain the fix to an upper version value.

[1] The FTP server that serves wcslib has only a few version back in
time, and actually doesn't serve wcslib 5.6 and 5.7 anymore. My
conclusion that this was introduced in 5.7 comes from inspecting the git
history of astropy, where they ported their bundled wcslib copy to 5.6
(commit 7bac470) and then to 5.7 (commit 3b3b11c).

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>