Fix memory leaks in the internal cache
The internal cache used by classes Euler and MVPosition allocates a number
of objects that are never deallocated. The leaked memory is fairly small,
but it adds a deluge of errors when code in casa is run under valgrind.

Note that the caching mechanism uses thread local storage (that has not been
changed). That means that any global object of type Euler or MVPosition could
lead to memory corruption and that there shouldn't be any sharing of instances
of those objects between threads.

Fixes CAS-10326