Remove altogether the Vector cache for MVPosition and Euler classes.
It was implemented using a thread_local storage, which means that
the cache shouldn't be shared among different threads. This was
unfortunely the case, since the MeasTable class creates some global
objects that include instances of MVPosition derived classes. This causes
undefined behaviours, since the constructors for thread_local variables are
called before the destructors for the global variables, according
to the C++11 standard. Running the tTableMeasures test triggered in fact
a segmentation fault due to this.

The original cache implementation didn't clean the memory, so the
problem was simply hidden. The original idea of the cache is
improve the performance, achieving around 5x boost when many objects
are allocated and deleted. However, in a quick test I replaced
Vector<Double> with std::array<double,3> (C++11) and I got a ~70x performance
boost without any extra cache, so that might be a better route.

Fixes CAS-10326