Supporting slice operations with Adios2StMan
This patch adds the initial support for cell slices in
Adios2StManColumn. This already existed i a small capacity with the
getSliceV method being already implemented, but: a) the corresponding
setSlice_ABC_V methods weren't implemneted, and b) this support wasn't
visible to the ArrayColumnT class because the canAccessSlice method
wasn't being overriden. These two shortcomings are addressed with this
patch, which makes the code compile, and the runtime happily write the
data.

Dimensions as understood by casacore are used in the reverse order by
the adios variable/selection mechanism as implemented in the Adios2StMan
(don't know why though). Thus the order in which we iterate over the
Slicer object must be the opposite we use to iterate over the adios
dimensionality structures.

We are also adding a std::string specialization that simply throws an
error stating that the method is not yet implemented.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>