Correct method name in error messages
Today while playing with the Adios2StMan I was surprised to see a
reference to putArray in my error messages, as I was actually juggling
with the putSlice* methods. I dug a bit and realized that the error
message was misleading. This patch fixes that, giving the correct method
name in the error message.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>