Using a pimpl idiom for Adios2StMan
The purpose of using this idiom is to hide the compile-time dependency on
ADIOS2 from users, who previously needed to have the correct -I pre-processor
flag when including the Adios2StMan.h header.

The impl class doesn't derive from DataManager like the
externally-facing class does. I initially tried with that inheritance in
place, but the software showed a runtime error that wasn't clear how to
connect to this inheritance. By removing the inheritance, and passing
down the externally-facing Adios2StMan object reference to the impl
class we therefore provide the actual implementation of the few methods
that are needed in impl (i.e., fileName() and ncolumns()).

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>