Allowing compilation against ADIOS 2.3.0
We are now using the CONFIG-based finding of ADIOS2 (as per b185392e4).
This is something ADIOS provides since 2.3.0, but the ADIOS2_LIBRARIES
variable was only re-defined back for backwards compatibility in 2.3.1,
leaving 2.3.0 as an unsuitable candidate to compile casacore against.

This commit changes the name we use to link against the ADIOS2 target,
which is always defined in the new CONFIG-based finding of ADIOS2.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>