Test showing faulty hashFunc behaviour
The intent of this hashFunc function is clearly to fill the individual bytes of
an uInt variable with various bytes via shifting and ORing together the
individual bytes. However the current implementation used a *logical* OR
operator. This meant that the four individual logical operations don't return
an integer value but rather either true or false, which then gets converted
into 1 or 0 when written into the uInt value, and then converted to true/false
again during the subsequent logical operations. From the four values involved
in the logical OR operations, only the sequential number can be zero, while the
rest will definitely be different from zero (when the object is "not null");
therefore it is expected that hashFunc always returns 1.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>