Feature addition in SPW and FIELD selection to use ID >= nRows for matching against name column of the sub-table.
If an integer ID in the selection expression is greater than or equal to the number of rows in the associated sub-tables, an attempt
is made match the ID as a string against the name column. If a match is found, the matched row number is selected.  Otherwise the usual
error handling happens. There should be no change for IDs less than the number of rows in the sub-table.