Improve CMake FindPython routines
CMake 3.12 introduced the FindPython2 and FindPython3 modules which
supercede the old and buggy FindPythonLibs and FindPythonInterp modules.
It finds consistent versions of the interpreter, development artifacts
and NumPy include directories in one go, and does it separately for
Python 2 and 3.

I've kept the original CMake variable names since I'm not sure where
they are used, but some of them may probably be removed or refactored.
This has been tested on a Homebrew setup (but not on Linux), where it
fixes detection errors using the previous commit.

By copying the appropriate modules from CMake, this functionality could
also be made to work on the older CMake 3.7 (although this might not be
old enough for all installations of casacore).