Pick Python CMakeLists.txt based on CMake version
Use the newer CMakeLists if CMake version >= 3.12. It would have been
nice to use the CMAKE_VERSION variable instead, but that only appeared
in CMake 2.6.3 (and we have a minimum version of 2.6.0).

An alternative solution is to check for the presence of the FindPython2
and FindPython3 modules, using a CheckHasModule function like the one
in https://stackoverflow.com/questions/15706318.