Improve CMake FindPython (#922)
* Rename original Python CMakeLists

These versions will be used by CMake 3.11 and older.

* Improve CMake FindPython routines

CMake 3.12 introduced the FindPython2 and FindPython3 modules which
supercede the old and buggy FindPythonLibs and FindPythonInterp modules.
It finds consistent versions of the interpreter, development artifacts
and NumPy include directories in one go, and does it separately for
Python 2 and 3.

I've kept the original CMake variable names since I'm not sure where
they are used, but some of them may probably be removed or refactored.
This has been tested on a Homebrew setup (but not on Linux), where it
fixes detection errors using the previous commit.

By copying the appropriate modules from CMake, this functionality could
also be made to work on the older CMake 3.7 (although this might not be
old enough for all installations of casacore).

* Pick Python CMakeLists.txt based on CMake version

Use the newer CMakeLists if CMake version >= 3.12. It would have been
nice to use the CMAKE_VERSION variable instead, but that only appeared
in CMake 2.6.3 (and we have a minimum version of 2.6.0).

An alternative solution is to check for the presence of the FindPython2
and FindPython3 modules, using a CheckHasModule function like the one
in https://stackoverflow.com/questions/15706318.

* Improve Boost.Python library name setting

Prefer EQUAL/GREATER to STREQUAL/STRGREATER, otherwise Boost 1.100
(Heaven forbid!) would be less than 1.67. Fix whitespace and reduce
the line count.