Use the same sort() call for the ms as for TableIterator.
The MSIter class was calling the sort() method of MeasurementSet
and then alling the constructor of TableIterator with the
NoSort option, assuming that the sorting was previously done.
However the sort parameters (columns, comparison
functions and sorting order) weren't the same one, leading
to wierd behaviour in cases where the comparison functions
or sorting order are not the default ones.

Additionally, a rounding error problem has been fixed in the case
of timeInterval = 0. In that case all timestamps should be
added to the same iteration group, but due to rounding errors
this wasn't always true.