Prevent setting OpenMP num_threads to zero
If a calculation of the number of threads previously yielded a value of zero,
replace it by a value of one.

AFAICT, the OpenMP specification states that a "num_threads" value must be a
positive integer or else the behavior is implementation defined. While the
meaning of "positive integer" appears undefined in the specification, it's usage
in that document suggests a value that is strictly greater than zero, which also
happens to agree with my intuitive sense of a minimum valid value of
"num_threads".

This bug originally appeared (for me) in Conversion.cc when using the OPARI2
tool (https://www.vi-hps.org/projects/score-p/) to instrument the casacore code,
while the bug apparently remains latent when using one of the standard compilers
without the tool.

Note that StatisticsDataset::initThreadVars() also computes a value for a number
of threads, and it does not check for zero. However, I have not changed that
code because I am uncertain that a value of zero would occur in practice, and
there is further initialization done in that method after computing the number
of threads.