Disable resetting the comparison function once the sorting already took place.
The comparison functions are used to sort the MS using the given
functions using Sort(). These comparison functions were also used
by next(). By allowing to change the comparison functions (in
particular MSIterval via setOffset() and setInterval()), an
inconsistency was created between sorting time and the iteration with
next().
Also a race condition has been removed in MSInterval regarding the
offset_p access. That member was modified the "first" time the comp()
function was called, but this is not deterministic in a parallel
algorithm like ParSort. Since setOffset() was also called inside
next() to "reset" the start of the time interval this imply a change
in the semantics of next().