Optimize the next() MSIter function by caching the sorting group boundaries.
The BaseTableIterator::next() method was doing an expensive walk
of the rows in determine where the next soring group was located.
However this information is already (almost) available when the
sorting takes place. With a call to sort::uniq() the sorting group
boundaries are already known. This patch propagted this information
in the iterator classes so that it can be used when iterating the ms.
The next() method will fallback to the old method if the iterator
has not been initialized with the proper group boundaries. This can
happen if the TableIterator::NoSort option is given and no group
boundaries are given.