Reverting reading cache and Table default engine
While working on the SUMMIT experiments, two set of changes were
introduced: first, a reading cache was added into Adios2StMan (e2038a464
and 1271e24f4), as it was believed that this would help with the
performance of the experiments. Additionally, the ADIOS2 engine to be
used by Adios2StMan was explicitly defaulted to "Table" instead of
relying on the default (c1e34d91e, fdad52b23, 735004b20 and 1a433559a).
These two changes were introduced with the aim of increasing the
performance of the storage manager, but in the end they were not used
during the parallel writing experiments, and yet already pretty good
writing performance was achieved (~130 GB/s with 1k writers).

This commit reverts these changes in order to prepare the code base for
introducing a separate set of changes to the storage manager that *did*
have a demonstrable impact on its performance. Note that the changes
being reverted here are not necessarily void of value, they have simply
not been demonstrated at scale. In time we could measure what their
effect is in parallel writing performance and maybe include them back,
but for the time being I would rather we focus on a single set of
changes.