Avoiding UB by not casting through char anymore
The previous version of this code stored the individual bytes of
information into a char variable, which was later shifted to place its
value into the final hash. This shifting was illegal though, as it took
an 8-bit variable type and shifted it by 8, 16 and 24 places to the
left. This might be the underlying reason causing #986, but even if it
isn't it still deserves to be fixed.

The fix avoids casting the individual bytes through the char type, and
instead it performs the masking and shifting in a single expression,
keeping the original value types and avoiding UB. Additionally, the last
byte is converted to uInt before bit shifting, as its original type is
char.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>