Remove support for CHANNEL_SELECTION keyword.
The CHANNEL_SELECTION keyword was used by CASA imager
to convey some information between cycles. This is not
used since time ago and it complicates unnecessarily the
logic in MSIter.
The startChan() method which delivers the first "selected"
channel is also removed, since there is no longer the concept
of channel selection.