Support explicit comparison functions for iteration grouping
A new constructor is created to support giving explicit comparison
functions to the table sorting algorithms. This makes MSIter more
generic. Eventually the other constructors could be removed if
all the code that uses MSIter (like VisibilityIterator* in CASA)
moves to the new API.
Note that this new constructor doesn't support cached SORTED_TABLE,
which could be considered legacy by now.

Additionally a a convenient sorting function which is always true
has been added. This comparison function could be convenient in cases where
one wants to ensure that all the values of a given column are
grouped together.

Unit tests have been added that exercise the new constructor.