Export metadata columns for each iteration
The array, field, data description metadata was assumed to be
constant in one given iteration. However, for applications that
need to get seveal SPWs or several fields in one single iteration
this is not true (those ones won't add those columns in the
sorting columns).

This change exports propagates the metadata columns for array, field,
dd so that the client can see the full metadata columns rather than
only the first value.

Additionally, the logic for the new*() methods has changed. If
the relevant column (or instance DataDescription) is not in the sorting
columns then one cannot be sure that all the elements in one iteration
have the same Id, therefore it is safer to simply signal that the
DataDescription is new.