Supports ParSort sorting algorithm in BaseTabIterator class
Currently, the TableIterator class has a parameter Option in
all the constructors which defaults to TableIterator::ParSort.
This parameter is propagated trough several classes:
TableIterator::TableIterator() -> BaseTable::makeIterator -> BaseTabIterator::BaseTabIterator()
However the later constructor BaseTabIterator::BaseTabIterator()
simply ignores the option ParSort (whereas QuickSort, HeapSort and
InsSort are supported.

This basically adds the support for ParSort as it is intended by
adding the proper "else if".