Make everything compile after Arrays changed
Most of the changes were necessary because Arrays pulls
no further headers in, and many files were dependent on this.
Besides headers, also lots of small changes because the
Array from Block constructor is gone, and because in
rare cases the new iterator constructors caused a
confusing resolve when the parameters were not matching.