Going for a different approach
Instead of dealing with the fact that the latest boost-python depends on
python@3.8, let's instead try to *not* update homebrew in the first
place -- this should speed things up considerably as "homebrew update"
won't be run anymore, many packages will not require re-installing,
dependencies won't be pulled, and so forth and so on. Hopefully this
will also mean that the boost-python and python package will *just*
work. If they don't I'll take it from there, but at least we'll have a
faster testing cycle.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>