Use FFTW instead of FFTPack in Fits-IDI to MS table converter
The FitsIDItoMS class is the only class that was still dependent on the FFTPack
Fortran code. By rewriting it to use FFTW, the FFTPack code can be removed.
This required implementing a wrapper for the real even FFT in the FFTW class,
as that was not available.

The FFTW class should be rewritten to have a more sensible interface. I've added
some TODO comments for that. Despite the use of a mutex, the FFTW wrapper is
also not thread safe.