Solve issue when assigning from lower to higher dim array
In psuedo code, this didn't work anymore, causing Casa errors:

Cube cube(5, 1, 1);
cube = Vector(5);

It did "by accident" work for a Matrix, because Matrix didn't implement a specialization of assignment, causing
an implicit call conversion via the Matrix(const Array&) constructor which did allow this.

This "feature" allowing assignment from lower dim arrays isn't yet documented and is also not consistent. For example,
this isn't allowed:

Array arr(IPosition{5, 1, 1});
arr = Vector(5);

So in the nearby future it would be good if i) this were documented; and ii) this would be made consistent
between Array and Matrix/Cube classes. By doing so, the specialized assignment methods from Matrix and Cube could
be removed, so it would also simplify the code.